requirements.txt
pandas>=1.3.0
numpy>=1.21.0
selenium>=4.0.0
webdriver-manager>=3.8.0
matplotlib>=3.5.0
seaborn>=0.11.0
python-dotenv>=0.19.0